requirements.txt
10xscale-agentflow[google-genai]>=0.7.2
PyQt6>=6.6
PyQt6-WebEngine>=6.6
qasync>=0.27
pydantic-settings>=2.0
python-dotenv>=1.0
pyyaml>=6.0
